Reworked Timers. Fixed bugs and new functionality (#33)
* Theoretically works, but needs to be tested

* Fix timer cancelling. Use vector to store timers that need to be cancelled, and loop through afterwards

* Did Colin's requested changes :/

* update data type for timer id usage

+++ include/robot/driveTest.h
@@ -16,7 +16,7 @@ private:
     float prevEncVelA;
     float prevEncVelB;
     float maxEncoderAccel = 0;
-    int encoderCheckTimerId;
+    unsigned long encoderCheckTimerId;
     bool testSuccessful;
     const int ENCODER_TOLERANCE = 100;
